requirements.txt
flask
flask-socketio
eventlet
python-dotenv
requests
psutil
pyserial
GPUtil
Pillow
